1870-06-10 Clarification of Report Priorities
The final report from the Board, issued on June 10th, 1870, reframed the criteria used when evaluating small arms. Contravening General Order No. 72, the Board stated that their investigation was conducted without regard to questions of caliber or ammunition. The Board's evaluation focused explicitly on the merits of the breech-loading action, and states that all other details of the experiments are considered incidental relative to this focus.